+++ code/game/ai_main.h
@@ -97,10 +97,6 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 #define PRESENCE_CROUCH				4
 //
 #define MAX_PROXMINES				64
-// For bot_developer mode
-#define BOTDEV_REPORTACTIONS		0x00000001
-#define BOTDEV_DISABLECHAT			0x00000002
-#define BOTDEV_TESTOBSTACLE			0x00000004
 // time the selected chat takes to type in
 #define BOTCHATTIME					2.0		//(float) trap_BotChatLength(bs->cs) * 30 / cpm;