+++ code/game/ai_vcmd.c
@@ -88,9 +88,10 @@ void BotVoiceChat_HoldPointA(bot_state_t *bs, int client, int mode) {
 	BotSetTeamStatus(bs);
 	// remember last ordered task
 	BotRememberLastOrderedTask(bs);
-#ifdef DEBUG
-	BotPrintTeamGoal(bs);
-#endif //DEBUG
+	// Developer mode outputs a message.
+	if (bot_developer.integer && bot_debugLTG.integer) {
+		BotPrintTeamGoal(bs);
+	}
 }
 
 /*
@@ -117,9 +118,10 @@ void BotVoiceChat_HoldPointB(bot_state_t *bs, int client, int mode) {
 	BotSetTeamStatus(bs);
 	// remember last ordered task
 	BotRememberLastOrderedTask(bs);
-#ifdef DEBUG
-	BotPrintTeamGoal(bs);
-#endif //DEBUG
+	// Developer mode outputs a message.
+	if (bot_developer.integer && bot_debugLTG.integer) {
+		BotPrintTeamGoal(bs);
+	}
 }
 
 /*
@@ -147,9 +149,10 @@ void BotVoiceChat_HoldDOMPoint(bot_state_t *bs, int client, int mode) {
 	BotSetTeamStatus(bs);
 	// remember last ordered task
 	BotRememberLastOrderedTask(bs);
-#ifdef DEBUG
-	BotPrintTeamGoal(bs);
-#endif //DEBUG
+	// Developer mode outputs a message.
+	if (bot_developer.integer && bot_debugLTG.integer) {
+		BotPrintTeamGoal(bs);
+	}
 }
 
 /*
@@ -189,7 +192,8 @@ void BotVoiceChat_GetFlag(bot_state_t *bs, int client, int mode) {
 	BotSetTeamStatus(bs);
 	// remember last ordered task
 	BotRememberLastOrderedTask(bs);
-	if (bot_developer.integer & BOTDEV_REPORTACTIONS) {
+	// Developer mode outputs a message.
+	if (bot_developer.integer && bot_debugLTG.integer) {
 		BotPrintTeamGoal(bs);
 	}
 }
@@ -263,7 +267,8 @@ void BotVoiceChat_Offense(bot_state_t *bs, int client, int mode) {
 		// remember last ordered task
 		BotRememberLastOrderedTask(bs);
 	}
-	if (bot_developer.integer & BOTDEV_REPORTACTIONS) {
+	// Developer mode outputs a message.
+	if (bot_developer.integer && bot_debugLTG.integer) {
 		BotPrintTeamGoal(bs);
 	}
 }
@@ -334,7 +339,8 @@ void BotVoiceChat_Defend(bot_state_t *bs, int client, int mode) {
 	BotSetTeamStatus(bs);
 	// remember last ordered task
 	BotRememberLastOrderedTask(bs);
-	if (bot_developer.integer & BOTDEV_REPORTACTIONS) {
+	// Developer mode outputs a message.
+	if (bot_developer.integer && bot_debugLTG.integer) {
 		BotPrintTeamGoal(bs);
 	}
 }
@@ -366,7 +372,8 @@ void BotVoiceChat_Patrol(bot_state_t *bs, int client, int mode) {
 	BotVoiceChatOnly(bs, -1, VOICECHAT_ONPATROL);
 	//
 	BotSetTeamStatus(bs);
-	if (bot_developer.integer & BOTDEV_REPORTACTIONS) {
+	// Developer mode outputs a message.
+	if (bot_developer.integer && bot_debugLTG.integer) {
 		BotPrintTeamGoal(bs);
 	}
 }
@@ -422,7 +429,8 @@ void BotVoiceChat_Camp(bot_state_t *bs, int client, int mode) {
 	BotSetTeamStatus(bs);
 	// remember last ordered task
 	BotRememberLastOrderedTask(bs);
-	if (bot_developer.integer & BOTDEV_REPORTACTIONS) {
+	// Developer mode outputs a message.
+	if (bot_developer.integer && bot_debugLTG.integer) {
 		BotPrintTeamGoal(bs);
 	}
 }
@@ -476,7 +484,8 @@ void BotVoiceChat_FollowMe(bot_state_t *bs, int client, int mode) {
 	BotSetTeamStatus(bs);
 	// remember last ordered task
 	BotRememberLastOrderedTask(bs);
-	if (bot_developer.integer & BOTDEV_REPORTACTIONS) {
+	// Developer mode outputs a message.
+	if (bot_developer.integer && bot_debugLTG.integer) {
 		BotPrintTeamGoal(bs);
 	}
 }
@@ -492,7 +501,8 @@ void BotVoiceChat_FollowFlagCarrier(bot_state_t *bs, int client, int mode) {
 	carrier = BotTeamFlagCarrier(bs);
 	if (carrier >= 0)
 		BotVoiceChat_FollowMe(bs, carrier, mode);
-	if (bot_developer.integer & BOTDEV_REPORTACTIONS) {
+	// Developer mode outputs a message.
+	if (bot_developer.integer && bot_debugLTG.integer) {
 		BotPrintTeamGoal(bs);
 	}
 }
@@ -519,7 +529,8 @@ void BotVoiceChat_ReturnFlag(bot_state_t *bs, int client, int mode) {
 	bs->teamgoal_time = FloatTime() + CTF_RETURNFLAG_TIME;
 	bs->rushbaseaway_time = 0;
 	BotSetTeamStatus(bs);
-	if (bot_developer.integer & BOTDEV_REPORTACTIONS) {
+	// Developer mode outputs a message.
+	if (bot_developer.integer && bot_debugLTG.integer) {
 		BotPrintTeamGoal(bs);
 	}
 }