+++ code/game/g_local.h
@@ -272,7 +272,7 @@ typedef struct {
 //unlagged - client options
 	// these correspond with variables in the userinfo string
 	int			delag;
-//	int			debugDelag;
+	int			debugDelag;
 	int			cmdTimeNudge;
 //unlagged - client options
 //unlagged - lag simulation #2